refactor: 계층형 구조 리팩토링 및 ProductOrder클래스 추가
기존 Product를 직접 거래하던 형식에서 주문 형식으로 교체하기 위한 리팩토링 진행
