+++ src/main/java/com/exception_study/api/user_account/controller/UserAccountController.java
@@ -1,0 +1,34 @@
+package com.exception_study.user_account.controller;
+
+import com.exception_study.global.dto.response.ResponseDto;
+import com.exception_study.user_account.service.UserAccountService;
+import com.exception_study.user_account.dto.request.LoginRequest;
+import com.exception_study.user_account.dto.response.LoginResponse;
+import com.exception_study.user_account.dto.response.SignUpResponse;
+import com.exception_study.user_account.dto.UserAccountDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/user")
+public class UserAccountController {
+    private final UserAccountService userAccountService;
+
+    @PostMapping("/signUp")
+    public ResponseDto<SignUpResponse> signUp(@RequestBody UserAccountDto dto){
+        SignUpResponse response = userAccountService.signUp(dto);
+        return ResponseDto.success(response);
+    }
+
+    @PostMapping("/login")
+    public ResponseDto<LoginResponse> login(@RequestBody LoginRequest dto){
+        LoginResponse response = userAccountService.login(dto);
+        return ResponseDto.success(response);
+    }
+
+
+}
