+++ src/main/java/com/exception_study/api/user_account/dto/request/LoginRequest.java
@@ -1,0 +1,9 @@
+package com.exception_study.user_account.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class LoginRequest {
+    private String userId;
+    private String password;
+}
