+++ src/main/java/com/exception_study/api/user_account/dto/request/RegisterRequest.java
@@ -1,0 +1,9 @@
+package com.exception_study.user_account.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class RegisterRequest {
+    private String name;
+    private int price;
+}
