+++ src/main/java/com/exception_study/api/user_account/dto/response/SignUpResponse.java
@@ -1,0 +1,15 @@
+package com.exception_study.user_account.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+@AllArgsConstructor
+@Data
+public class SignUpResponse {
+    private String message;
+
+
+    public static SignUpResponse of(String message){
+        return new SignUpResponse(message);
+    }
+}
+
