+++ src/main/java/com/exception_study/api/user_account/repository/UserAccountRepository.java
@@ -1,0 +1,7 @@
+package com.exception_study.user_account.repository;
+
+import com.exception_study.user_account.entity.UserAccount;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserAccountRepository extends JpaRepository<UserAccount,String> {
+}
