--- src/main/java/com/exception_study/dto/UserAccountDto.java
@@ -1,10 +1,0 @@
-package com.exception_study.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserAccountDto {
-    private String userId;
-    private String password;
-    private String userName;
-}
