--- src/main/java/com/exception_study/dto/request/LoginRequest.java
@@ -1,9 +1,0 @@
-package com.exception_study.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequest {
-    private String userId;
-    private String password;
-}
