--- src/main/java/com/exception_study/dto/request/RegisterRequest.java
@@ -1,9 +1,0 @@
-package com.exception_study.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class RegisterRequest {
-    private String name;
-    private int price;
-}
