--- src/main/java/com/exception_study/dto/response/LoginResponse.java
@@ -1,16 +1,0 @@
-package com.exception_study.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class LoginResponse {
-    private String token;
-    private long expiredTime;
-
-    public static LoginResponse of(String token, long expiredTime){
-        return new LoginResponse(token,expiredTime);
-    }
-}
