--- src/main/java/com/exception_study/exception/GlobalControllerAdvice.java
@@ -1,29 +1,0 @@
-package com.exception_study.exception;
-
-import com.exception_study.dto.response.ResponseDto;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import static com.exception_study.exception.ErrorCode.DATABASE_ERROR;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalControllerAdvice {
-
-    @ExceptionHandler(StudyApplicationException.class)
-    public ResponseEntity<?> errorHandler(StudyApplicationException e){
-        log.error("Error occurs {}", e.toString());
-        return ResponseEntity.status(e.getErrorCode().getStatus())
-                .body(ResponseDto.error(e.getErrorCode().name()));
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<?> databaseErrorHandler(IllegalArgumentException e) {
-        log.error("Error occurs {}", e.toString());
-        return ResponseEntity.status(DATABASE_ERROR.getStatus())
-                .body(ResponseDto.error(DATABASE_ERROR.name()));
-    }
-
-}
