--- src/main/java/com/exception_study/repository/UserAccountRepository.java
@@ -1,7 +1,0 @@
-package com.exception_study.repository;
-
-import com.exception_study.entity.UserAccount;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserAccountRepository extends JpaRepository<UserAccount,String> {
-}
