+++ src/main/resources/data.sql
@@ -1,8 +1,19 @@
-# insert into product(id, name, price, status, seller, buyer) values (1,'블루투스 이어폰',100000,'판매중','테스터1',null),
-#        (2,'모자',20000,'예약중','테스터2','오두호'),
-#        (3,'볼펜',1000,'예약중','테스터3','테스터4'),
-#        (4,'노트북',1000000,'판매중','오두호',null),
-#        (5,'블루투스 스피커',50000,'완료','오두호','테스터5');
+# drop table if exists user_account;
+#
+# create table user_account(
+#     user_id varchar(255) primary key ,
+#     password varchar(255),
+#     user_name varchar(255)
+# );
+
+
+
+
+# insert into product(id, name, price, status, seller, buyer) values (1,'블루투스 이어폰',100000,'판매중','테스터1'),
+#        (2,'모자',20000,'예약중','테스터2'),
+#        (3,'볼펜',1000,'예약중','테스터3'),
+#        (4,'노트북',1000000,'판매중','오두호'),
+#        (5,'블루투스 스피커',50000,'완료','오두호');
 #
